+++ src/rcore.c
@@ -3148,8 +3148,8 @@ const char *GetApplicationDirectory(void)
 
     if (_NSGetExecutablePath(appDir, &size) == 0)
     {
-        int len = strlen(appDir);
-        for (int i = len; i >= 0; --i)
+        long len = strlen(appDir);
+        for (long i = len; i >= 0; --i)
         {
             if (appDir[i] == '/')
             {
