+++ src/utils.c
@@ -200,7 +200,7 @@ unsigned char *LoadFileData(const char *fileName, unsigned int *bytesRead)
             // WARNING: On binary streams SEEK_END could not be found,
             // using fseek() and ftell() could not work in some (rare) cases
             fseek(file, 0, SEEK_END);
-            int size = ftell(file);
+            size_t size = ftell(file);
             fseek(file, 0, SEEK_SET);
 
             if (size > 0)
